BUG: Somehow the scale is messed up with marm2 example
Issue is that the waves are not actually making it to the receiver locations.
We need to change how long we run the simulation, etc. based on how long we expect such a signal to reach.
Note that the issue could also be that the source amplitudes are too weak, and the waves are actually arriving but are weak.
This is not as likely as that the waves don't have enough time to arrive, but it should be kept in mind.