MILESTONE: THIS IS IT THIS IS IT
This is the example you need
MEGA MILESTONE

It's clear as day.

Just comment the one out versus the other.

Note that you need to let sigma be sufficiently large for the gradients to flow.
But as long as you do that quasi-W1 gets the type of convergence we want.